[#44] Fix : 이메일 의존성 주입 불가 문제 해결
