+++ src/main/java/com/b3/ddarelro/domain/user/controller/EmailController.java
@@ -4,7 +4,7 @@ import com.b3.ddarelro.domain.user.dto.request.EmailAuthSendReq;
 import com.b3.ddarelro.domain.user.dto.request.EmailCodeCheckReq;
 import com.b3.ddarelro.domain.user.dto.response.EmailAuthSendRes;
 import com.b3.ddarelro.domain.user.dto.response.EmailCodeCheckRes;
-import com.b3.ddarelro.domain.user.service.EmailAuthService;
+import com.b3.ddarelro.domain.user.service.EmailService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -17,14 +17,14 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/api/email")
-public class EmailAuthController {
+public class EmailController {
 
-    private final EmailAuthService emailAuthService;
+    private final EmailService emailService;
 
     @PostMapping
     public ResponseEntity<EmailAuthSendRes> sendEmail(
         @Valid @RequestBody EmailAuthSendReq reqDto) {
-        EmailAuthSendRes resDto = emailAuthService.sendEmail(reqDto);
+        EmailAuthSendRes resDto = emailService.sendEmail(reqDto);
         return ResponseEntity.ok(resDto);
     }
 
@@ -32,7 +32,7 @@ public class EmailAuthController {
     public ResponseEntity<EmailCodeCheckRes> checkAuthCode(
         @Valid @RequestBody EmailCodeCheckReq reqDto) {
 
-        EmailCodeCheckRes resDto = emailAuthService.checkAuthCode(reqDto);
+        EmailCodeCheckRes resDto = emailService.checkAuthCode(reqDto);
         return ResponseEntity.ok(resDto);
     }
